# models/review.js
const { Sequelize, DataTypes } = require('sequelize');
const { sequelizePostgres } = require('../dataBase/database');

const Review = sequelizePostgres.define('Review', {
    id: {
      type: DataTypes.INTEGER,
      autoIncrement: true,
      primaryKey: true
    },
    movieId: {
      type: DataTypes.INTEGER,
      allowNull: false,
      references: {
        model: 'Movies',
        key: 'id'
      }
    },
    userId: {
      type: DataTypes.INTEGER,
      allowNull: false,
      references: {
        model: 'Users',
        key: 'id'
      }
    },
    message: {
      type: DataTypes.TEXT,
      allowNull: true
    },
    post_date:{
        type: DataTypes.DATE,
        allowNull: false
    },
   like:{
    type: DataTypes.INTEGER,
    allowNull: true,
    defaultValue: 0
   }
  });
  
  module.exports = Review;
  